stats-service/requirements.txt
fastapi[standard]==0.115.12
sqlalchemy[asyncio]==2.0.41
aio-pika==9.5.5
fastapi-cache2[redis]==0.2.2
asyncpg==0.30.0
prometheus-fastapi-instrumentator==7.1.0